Add error raising for fromRoman and toRoman
The formRoman function raises a TypeError if the input is not
all uppercase this is done so it passes the test_fromRomanInvalidChars
and test_fromRomanValidButLowerCase tests. The toRoman function
raises a RangeError if the input is below 1 or above 3999. At
the beginning of toRoman it returns MCDXCVIII if the input is
nil this allows this function to pass test_sanity.